still needs testing and  editing (but to be used with forward chaining)

+++ Source/InferenceEngine/Core/Response.hpp
@@ -12,8 +12,8 @@
 class Response {
 public:
     bool Result = false;
-    int ModalsInfered = 0;
-    Response(int mods, bool res) : ModalsInfered(mods), Result(res) {};
+    double ModalsInfered = 0;
+    Response(double mods, bool res) : ModalsInfered(mods), Result(res) {};
 };
